+++ packages/smath/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.2.0
+
+- Add `gcd()` to compute the greatest common denominator of two integers
+- Determine if numbers are integers using `n % 1 !== 0`
+- Add example to `isPrime()`
+- Add tests for `isPrime()` and `gcd()`
+
 ## 2.1.0
 
 - Add `isPrime()` to check if any number is prime